fix: Change component names to `daikin-accordion` and `daikin-accordion-item`
