--- src/components/accordion-group/daikin-accordion-group.ts
@@ -1,39 +1,0 @@
-import { cva } from "class-variance-authority";
-import { LitElement, css, html, unsafeCSS } from "lit";
-import { customElement } from "lit/decorators.js";
-import tailwindStyles from "../../tailwind.css?inline";
-
-const cvaContainer = cva(["w-full", "border-y-[1px]", "border-y-[#CECECE]"]);
-
-/**
- * Primary UI component for user interaction
- */
-@customElement("daikin-accordion-group")
-export class DaikinAccordionGroup extends LitElement {
-  static override readonly styles = css`
-    ${unsafeCSS(tailwindStyles)}
-
-    :host {
-      display: block;
-      width: 100%;
-      max-width: 400px;
-    }
-    ::slotted(daikin-accordion:not(:last-child)) {
-      border-bottom: 1px solid #cecece;
-    }
-  `;
-
-  override render() {
-    return html`<div class=${cvaContainer()}>
-      <slot></slot>
-    </div>`;
-  }
-}
-
-declare global {
-  interface HTMLElementTagNameMap {
-    "daikin-accordion-group": DaikinAccordionGroup;
-  }
-}
-
-export default DaikinAccordionGroup;
