--- src/components/accordion-group/index.ts
@@ -1,1 +1,0 @@
-export * from "./daikin-accordion-group";
