+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/close-disabled-base-1.png
