+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/close-disabled-focus-1.png
