+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/close-disabled-hover-1.png
