+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/close-disabled-press-1.png
