+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/close-enabled-base-1.png
