+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/close-enabled-focus-1.png
