+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/close-enabled-hover-1.png
