+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/close-enabled-press-1.png
