+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/open-disabled-base-1.png
