+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/open-disabled-focus-1.png
