+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/open-disabled-hover-1.png
