+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/open-disabled-press-1.png
