+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/open-enabled-base-1.png
