+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/open-enabled-focus-1.png
