+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/open-enabled-hover-1.png
