+++ src/components/accordion-item/daikin-accordion-item.visual.test.ts-snapshots/open-enabled-press-1.png
