+++ src/components/accordion-item/index.ts
@@ -1,0 +1,1 @@
+export * from "./daikin-accordion-item";
