+++ src/components/accordion/daikin-accordion.visual.test.ts-snapshots/base-1.png
